Improve package split
