+++ src/YiisoftFormModel/FormModel.php
@@ -7,11 +7,11 @@ namespace Yiisoft\Form\YiisoftFormModel;
 use InvalidArgumentException;
 use ReflectionClass;
 use ReflectionException;
-use Yiisoft\Form\Exception\PropertyNotSupportNestedValuesException;
-use Yiisoft\Form\Exception\StaticObjectPropertyException;
-use Yiisoft\Form\Exception\UndefinedArrayElementException;
-use Yiisoft\Form\Exception\UndefinedObjectPropertyException;
-use Yiisoft\Form\Exception\ValueNotFoundException;
+use Yiisoft\Form\YiisoftFormModel\Exception\PropertyNotSupportNestedValuesException;
+use Yiisoft\Form\YiisoftFormModel\Exception\StaticObjectPropertyException;
+use Yiisoft\Form\YiisoftFormModel\Exception\UndefinedArrayElementException;
+use Yiisoft\Form\YiisoftFormModel\Exception\UndefinedObjectPropertyException;
+use Yiisoft\Form\YiisoftFormModel\Exception\ValueNotFoundException;
 use Yiisoft\Hydrator\Validator\ValidatedInputTrait;
 use Yiisoft\Strings\Inflector;
 use Yiisoft\Strings\StringHelper;
