+++ tests/YiisoftFormModel/Support/Form/DynamicAttributesForm.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Form\Tests\TestSupport\Form;
+namespace Yiisoft\Form\Tests\YiisoftFormModel\Support\Form;
 
 use Yiisoft\Arrays\ArrayHelper;
 use Yiisoft\Form\YiisoftFormModel\FormModel;
